+++ src/extension.ts
@@ -149,9 +149,9 @@ async function activate(context: vscode.ExtensionContext) {
     context.subscriptions.push(vscode.commands.registerCommand('calva.startStandaloneRepl', () => {
         replStart.startStandaloneRepl(context, replStart.USER_TEMPLATE_FILE_NAMES, true);
     }));
-    //context.subscriptions.push(vscode.commands.registerCommand('calva.startStandaloneHelloRepl', () => {
-    //    replStart.startStandaloneRepl(context, replStart.HELLO_TEMPLATE_FILE_NAMES, false);
-    //}));
+    context.subscriptions.push(vscode.commands.registerCommand('calva.startStandaloneHelloRepl', () => {
+        replStart.startStandaloneRepl(context, replStart.HELLO_TEMPLATE_FILE_NAMES, false);
+    }));
     context.subscriptions.push(vscode.commands.registerCommand('calva.jackIn', jackIn.jackInCommand));
     context.subscriptions.push(vscode.commands.registerCommand('calva.copyJackInCommandToClipboard', jackIn.copyJackInCommandToClipboard));
     context.subscriptions.push(vscode.commands.registerCommand('calva.connectNonProjectREPL', () => {
@@ -168,6 +168,7 @@ async function activate(context: vscode.ExtensionContext) {
     }));
     context.subscriptions.push(vscode.commands.registerCommand('calva.interruptAllEvaluations', eval.interruptAllEvaluations));
     context.subscriptions.push(vscode.commands.registerCommand('calva.evaluateSelection', eval.evaluateCurrentForm));
+    context.subscriptions.push(vscode.commands.registerCommand('calva.evaluateToCursor', eval.evaluateToCursor));
     context.subscriptions.push(vscode.commands.registerCommand('calva.evaluateUser', eval.evaluateUser));
     context.subscriptions.push(vscode.commands.registerCommand('calva.evaluateCurrentTopLevelForm', eval.evaluateTopLevelForm));
     context.subscriptions.push(vscode.commands.registerCommand('calva.evaluateSelectionReplace', eval.evaluateSelectionReplace));